Add StudentAdditionalDetail test model
[Closes #23]

+++ app/models/student_additional_detail.rb
@@ -19,30 +19,21 @@
 class StudentAdditionalDetail < ActiveRecord::Base
   belongs_to :student
   belongs_to :student_additional_field, :foreign_key=>'additional_field_id'
-  #validates_presence_of :additional_info
+  validates_presence_of :additional_info, :if => :student_additional_field_is_mandatory?
+  before_validation :destroy_when_additional_info_blank
 
   def save
-    unless self.destroyed?
-      super
-    end
+    super unless self.destroyed?
     true
   end
 
-  def validate
-    if self.student_additional_field.is_mandatory == true
-      if self.additional_info.blank?
-        errors.add("additional_info","can't be blank")
-        return false
-      else
-        return true
-      end
-    else
-      if self.additional_info.blank?
-        self.destroy
-        return true
-      else
-        return true
-      end
-    end
+  def student_additional_field_is_mandatory?
+    student_additional_field && student_additional_field.is_mandatory?
+  end
+
+  private
+
+  def destroy_when_additional_info_blank
+    self.destroy if self.additional_info.blank? || !student_additional_field_is_mandatory?
   end
 end
